Add KoinTestFixture
