+++ core/test-util/src/main/kotlin/edu/wpi/axon/testutil/KoinTestFixture.kt
@@ -1,0 +1,20 @@
+package edu.wpi.axon.testutil
+
+import org.junit.jupiter.api.AfterEach
+import org.koin.core.context.stopKoin
+import org.koin.test.KoinTest
+
+open class KoinTestFixture : KoinTest {
+
+    private var additionalAfterEach: () -> Unit = {}
+
+    @AfterEach
+    fun afterEach() {
+        additionalAfterEach()
+        stopKoin()
+    }
+
+    fun additionalAfterEach(configure: () -> Unit) {
+        additionalAfterEach = configure
+    }
+}
